requirements.txt
python-dotenv>=1.0
requests>=2.31
prompt-toolkit>=3.0
rich>=13.7
